Refactored the massive singleboat component. Created a better mobile navbar. Minor style changes.
