+++ frontend/public/index.html
@@ -22,7 +22,7 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="A social media app for sharing boat trips and maintenance"
     />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>WeSail</title>
